src/requirements.txt
python-dateutil
icalendar
pytz
requests
Flask
supervisor
